Cut from new sji to sji
